Changes call rates from decimal to float.

+++ src/cgr_gwas_qc/models/software_params.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from pydantic import BaseModel, Field, validator
 
 
@@ -8,10 +6,10 @@ class SoftwareParams(BaseModel):
         "TOP",
         description="Which strand to use when converting GTC to plink {TOP, FWD, PLUS}. default TOP",
     )
-    samp_cr_1: Decimal = Field(0.8, description="Sample call rate filter 1", gt=0, le=1)
-    snp_cr_1: Decimal = Field(0.8, description="Marker call rate filter 1", gt=0, le=1)
-    samp_cr_2: Decimal = Field(0.95, description="Sample call rate filter 2", gt=0, le=1)
-    snp_cr_2: Decimal = Field(0.95, description="Marker call rate filter 2", gt=0, le=1)
+    samp_cr_1: float = Field(0.8, description="Sample call rate filter 1", gt=0, le=1)
+    snp_cr_1: float = Field(0.8, description="Marker call rate filter 1", gt=0, le=1)
+    samp_cr_2: float = Field(0.95, description="Sample call rate filter 2", gt=0, le=1)
+    snp_cr_2: float = Field(0.95, description="Marker call rate filter 2", gt=0, le=1)
 
     @validator("strand")
     def validate_strand(cls, v):
